# source/Spydersoft.Identity.Admin.Frontend/admin-ui/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Spydersoft.Identity.Admin.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:7030/"
    }
  ],
  "paths": {
...
    "/api/v1/Clients/{id}": {
      "get": {
        "tags": [
          "Clients"
        ],
        "summary": "Returns the full detail of a single client.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ClientDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Clients"
        ],
        "summary": "Updates an existing client.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveClientDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveClientDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SaveClientDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Clients"
        ],
        "summary": "Deletes a client.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClientDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "clientUri": {
            "type": [
              "null",
              "string"
            ]
          },
          "logoUri": {
            "type": [
              "null",
              "string"
            ]
          },
          "requireClientSecret": {
            "type": "boolean"
          },
          "requireConsent": {
            "type": "boolean"
          },
          "requirePkce": {
            "type": "boolean"
          },
          "allowPlainTextPkce": {
            "type": "boolean"
          },
          "allowOfflineAccess": {
            "type": "boolean"
          },
          "allowAccessTokensViaBrowser": {
            "type": "boolean"
          },
          "allowRememberConsent": {
            "type": "boolean"
          },
          "alwaysSendClientClaims": {
            "type": "boolean"
          },
          "alwaysIncludeUserClaimsInIdToken": {
            "type": "boolean"
          },
          "enableLocalLogin": {
            "type": "boolean"
          },
          "includeJwtId": {
            "type": "boolean"
          },
          "nonEditable": {
            "type": "boolean"
          },
          "updateAccessTokenClaimsOnRefresh": {
            "type": "boolean"
          },
          "backChannelLogoutSessionRequired": {
            "type": "boolean"
          },
          "frontChannelLogoutSessionRequired": {
            "type": "boolean"
          },
          "absoluteRefreshTokenLifetime": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "accessTokenLifetime": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "accessTokenType": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "authorizationCodeLifetime": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "deviceCodeLifetime": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "identityTokenLifetime": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "refreshTokenExpiration": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "refreshTokenUsage": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "slidingRefreshTokenLifetime": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "consentLifetime": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "userSsoLifetime": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "backChannelLogoutUri": {
            "type": [
              "null",
              "string"
            ]
          },
          "clientClaimsPrefix": {
            "type": [
              "null",
              "string"
            ]
          },
          "frontChannelLogoutUri": {
            "type": [
              "null",
              "string"
            ]
          },
          "pairWiseSubjectSalt": {
            "type": [
              "null",
              "string"
            ]
          },
          "userCodeType": {
            "type": [
              "null",
              "string"
            ]
          },
          "updated": {
            "type": "string",
            "format": "date-time"
          },
          "lastAccessed": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "Gets or sets the identifier.",
            "format": "int32"
          },
          "clientId": {
            "type": "string",
            "description": "Gets or sets the client identifier string."
          },
          "clientName": {
            "type": [
              "null",
              "string"
            ],
            "description": "Gets or sets the display name."
          },
          "enabled": {
            "type": "boolean",
            "description": "Gets or sets a value indicating whether this client is enabled."
          },
          "protocolType": {
            "type": "string",
            "description": "Gets or sets the protocol type."
          },
          "created": {
            "type": "string",
            "description": "Gets or sets the creation timestamp.",
            "format": "date-time"
          }
        },
        "description": "Full client detail DTO."
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "SaveClientDto": {
        "required": [
          "clientId",
          "clientName",
          "protocolType"
        ],
        "type": "object",
        "properties": {
          "clientId": {
            "maxLength": 200,
            "minLength": 2,
            "type": "string"
          },
          "clientName": {
            "maxLength": 200,
            "minLength": 2,
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "clientUri": {
            "type": [
              "null",
              "string"
            ]
          },
          "logoUri": {
            "type": [
              "null",
              "string"
            ]
          },
          "requireClientSecret": {
            "type": "boolean"
          },
          "requireConsent": {
            "type": "boolean"
          },
          "requirePkce": {
            "type": "boolean"
          },
          "allowPlainTextPkce": {
            "type": "boolean"
          },
          "allowOfflineAccess": {
            "type": "boolean"
          },
          "allowAccessTokensViaBrowser": {
            "type": "boolean"
          },
          "allowRememberConsent": {
            "type": "boolean"
          },
          "alwaysSendClientClaims": {
            "type": "boolean"
          },
          "alwaysIncludeUserClaimsInIdToken": {
            "type": "boolean"
          },
          "enableLocalLogin": {
            "type": "boolean"
          },
          "includeJwtId": {
            "type": "boolean"
          },
          "nonEditable": {
            "type": "boolean"
          },
          "updateAccessTokenClaimsOnRefresh": {
            "type": "boolean"
          },
          "backChannelLogoutSessionRequired": {
            "type": "boolean"
          },
          "frontChannelLogoutSessionRequired": {
            "type": "boolean"
          },
          "enabled": {
            "type": "boolean"
          },
          "absoluteRefreshTokenLifetime": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "accessTokenLifetime": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "accessTokenType": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "authorizationCodeLifetime": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "deviceCodeLifetime": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "identityTokenLifetime": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "refreshTokenExpiration": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "refreshTokenUsage": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "slidingRefreshTokenLifetime": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "consentLifetime": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "userSsoLifetime": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "backChannelLogoutUri": {
            "type": [
              "null",
              "string"
            ]
          },
          "clientClaimsPrefix": {
            "type": [
              "null",
              "string"
            ]
          },
          "frontChannelLogoutUri": {
            "type": [
              "null",
              "string"
            ]
          },
          "pairWiseSubjectSalt": {
            "type": [
              "null",
              "string"
            ]
          },
          "protocolType": {
            "maxLength": 200,
            "minLength": 2,
            "type": "string"
          },
          "userCodeType": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "description": "Payload for creating or updating a client."
      },
...
    }
  },
...
}